+++ src/main/java/com/example/shortform/mail/HtmlEmailService.java
@@ -25,7 +25,7 @@ public class HtmlEmailService implements EmailService {
             MimeMessageHelper mimeMessageHelper = new MimeMessageHelper(mimeMessage, false, "UTF-8");
             mimeMessageHelper.setTo(emailMessage.getTo());
             mimeMessageHelper.setSubject(emailMessage.getSubject());
-            mimeMessageHelper.setText(emailMessage.getMessage(), true);
+            mimeMessageHelper.setText(emailMessage.getMessage(), false);
             javaMailSender.send(mimeMessage);
 
             log.info("sent email: {}", emailMessage.getMessage());
